requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
cryptography>=42.0.4
firebase-admin>=6.2.0
passlib[bcrypt]>=1.7.4